refactor:입력필드 및 submit 버튼 shared로 이동
