+++ src/shared/ui/AuthLayout.tsx
