+++ src/shared/ui/index.ts
@@ -1,1 +1,3 @@
-export { default as TextField } from '../../features/auth-login/ui/TextField';
+export { default as AuthLayout } from './AuthLayout';
+export { default as TextField } from './TextField';
+export { default as SubmitButton } from './SubmitButton';
