--- src/widgets/login/ui/index.ts
@@ -1,5 +1,0 @@
-// Auth UI 세그먼트 Public API
-// export * from './LoginForm';
-// export * from './SignupForm';
-// export * from './LogoutButton';
-export { default as LoginForm } from './LoginForm';
